Revert "Add pocket base as db and auth"
