--- .env
@@ -1,3 +1,0 @@
-VITE_POCKET_BASE_URL='http://0.0.0.0:8090'
-
-// VITE_POCKET_BASE_URL='https://dnd-helper.pockethost.io'
