+++ src/SpellIcon.ts
@@ -1,5 +1,5 @@
-import { Spell } from '@/types/spell';
 import FireBall from '@/assets/fire-ball.png';
+import { Spell } from '@/SpellList';
 
 const getImgUrl = (name: string): string => {
   switch (Spell[name]) {
