+++ src/components/SideMenu.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
   import useAppStore, { Route } from '@/stores/appStore';
-  import LoginForm from '@/components/LoginForm.vue';
 
   withDefaults(
     defineProps<{
@@ -32,7 +31,6 @@
     <div @click="openMySpells">Книга заклять</div>
     <div @click="openAllSpells">Всі закляття</div>
     <div @click="openSkills">Вміння</div>
-    <LoginForm />
   </div>
 </template>
