+++ src/stores/appStore.ts
@@ -1,7 +1,7 @@
 import { defineStore } from 'pinia';
+
 import { Ref, ref } from 'vue';
-import { Spell } from '@/types/spell';
-import PocketDataBase from '@/api/PocketBase';
+import { Spell } from '@/SpellList';
 
 export enum Route {
   MySpells,
@@ -9,8 +9,6 @@ export enum Route {
   Skills,
 }
 const useAppStore = defineStore('app', () => {
-  const db = new PocketDataBase();
-
   const route: Ref<Route> = ref(Route.MySpells);
   const menuIsOpen: Ref<boolean> = ref(false);
   const selectedSpells: Ref<Array<Spell>> = ref([]);
@@ -33,7 +31,6 @@ const useAppStore = defineStore('app', () => {
   };
 
   return {
-    db,
     route,
 
     menuIsOpen,
